+++ tests/test_cloud_run_worker_v2.py
@@ -2,7 +2,7 @@ import pytest
 from prefect.utilities.dockerutils import get_prefect_image_name
 
 from prefect_gcp.credentials import GcpCredentials
-from prefect_gcp.worker_v2 import CloudRunWorkerJobV2Configuration
+from prefect_gcp.workers.worker_v2 import CloudRunWorkerJobV2Configuration
 
 
 @pytest.fixture